Refactored where components live and how they are imported
